Icons and 'more' button fix
I removed the container around the menu icons and the need for centering and extra stroke disappeared. Thanks to this, hover and focus are applied to the icons themselves. Added transparent borders to elements so they don't jump when focusing on them. The "more" button was removed from the grid and the display block was used - it became the entire width of the popup. I put the second part of the pop-up in a separate container. Added a border between the initial content and additional content. Slightly reworked the indents.